+++ packages/pi-conductor/__tests__/static-safety.test.ts
@@ -75,9 +75,23 @@ describe("pi-conductor static safety guards", () => {
     expect(storageSource.split("\n").length).toBeLessThan(1600);
     expect(storageSource).toContain('from "./storage-validation.js"');
     expect(storageSource).toContain('from "./storage-normalize.js"');
+    expect(storageSource).toContain('from "./storage-query.js"');
     expect(storageSource).not.toContain("conductorEventTypes");
     expect(validationSource.split("\n").length).toBeLessThan(420);
     expect(normalizeSource.split("\n").length).toBeLessThan(120);
+  });
+
+  it("keeps storage query pagination in a dedicated query module", () => {
+    const extensionDir = join(__dirname, "../extensions");
+    const storageSource = readFileSync(join(extensionDir, "storage.ts"), "utf-8");
+    const querySource = readFileSync(join(extensionDir, "storage-query.ts"), "utf-8");
+
+    expect(storageSource).toContain('from "./storage-query.js"');
+    expect(storageSource).not.toContain("export function queryConductorArtifacts");
+    expect(storageSource).not.toContain("export function queryConductorEvents");
+    expect(querySource).toContain("export function queryConductorArtifacts");
+    expect(querySource).toContain("export function queryConductorEvents");
+    expect(querySource.split("\n").length).toBeLessThan(120);
   });
 
   it("keeps next-action recommendation planning in a pure conductor module", () => {
@@ -119,6 +133,62 @@ describe("pi-conductor static safety guards", () => {
     expect(objectiveSource).toContain("export function planObjectiveForRepo");
     expect(objectiveSource.split("\n").length).toBeLessThan(240);
     expect(repoRunSource.split("\n").length).toBeLessThan(80);
+  });
+
+  it("keeps evidence bundles and human review packets outside the conductor orchestrator", () => {
+    const extensionDir = join(__dirname, "../extensions");
+    const conductorSource = readFileSync(join(extensionDir, "conductor.ts"), "utf-8");
+    const evidenceSource = readFileSync(join(extensionDir, "evidence-service.ts"), "utf-8");
+    const reviewSource = readFileSync(join(extensionDir, "review-service.ts"), "utf-8");
+
+    expect(conductorSource).toContain('from "./evidence-service.js"');
+    expect(conductorSource).toContain('from "./review-service.js"');
+    expect(conductorSource).not.toContain("function isTerminalStatus");
+    expect(conductorSource).not.toContain("export function buildEvidenceBundleForRepo");
+    expect(conductorSource).not.toContain("export function checkReadinessForRepo");
+    expect(conductorSource).not.toContain("export function buildBlockingDiagnosisForRepo");
+    expect(conductorSource).not.toContain("export function prepareHumanReviewForRepo");
+    expect(evidenceSource).toContain("export function buildEvidenceBundleForRepo");
+    expect(evidenceSource).toContain("export function checkReadinessForRepo");
+    expect(evidenceSource.split("\n").length).toBeLessThan(260);
+    expect(reviewSource).toContain("export function buildBlockingDiagnosisForRepo");
+    expect(reviewSource).toContain("export function prepareHumanReviewForRepo");
+    expect(reviewSource.split("\n").length).toBeLessThan(140);
+  });
+
+  it("keeps task lifecycle mutation outside the conductor orchestrator", () => {
+    const extensionDir = join(__dirname, "../extensions");
+    const conductorSource = readFileSync(join(extensionDir, "conductor.ts"), "utf-8");
+    const taskSource = readFileSync(join(extensionDir, "task-service.ts"), "utf-8");
+
+    expect(conductorSource).toContain('from "./task-service.js"');
+    expect(conductorSource).not.toContain("function createTaskId");
+    expect(conductorSource).not.toContain("function createRunId");
+    expect(conductorSource).not.toContain("export function createTaskForRepo");
+    expect(conductorSource).not.toContain("export function recordTaskCompletionForRepo");
+    expect(conductorSource).not.toContain("export function startTaskRunForRepo");
+    expect(conductorSource).not.toContain("export function createFollowUpTaskForRepo");
+    expect(taskSource).toContain("export function createTaskForRepo");
+    expect(taskSource).toContain("export function recordTaskCompletionForRepo");
+    expect(taskSource).toContain("export function startTaskRunForRepo");
+    expect(taskSource).toContain("refreshObjectiveStatusForRepo");
+    expect(taskSource.split("\n").length).toBeLessThan(320);
+  });
+
+  it("keeps gate mutation outside the conductor orchestrator", () => {
+    const extensionDir = join(__dirname, "../extensions");
+    const conductorSource = readFileSync(join(extensionDir, "conductor.ts"), "utf-8");
+    const gateSource = readFileSync(join(extensionDir, "gate-service.ts"), "utf-8");
+
+    expect(conductorSource).toContain('from "./gate-service.js"');
+    expect(conductorSource).not.toContain("function createGateId");
+    expect(conductorSource).not.toContain("export function createGateForRepo");
+    expect(conductorSource).not.toContain("export function resolveGateForRepo");
+    expect(conductorSource).not.toContain("export function resolveGateFromTrustedHumanForRepo");
+    expect(gateSource).toContain("export function createGateForRepo");
+    expect(gateSource).toContain("export function resolveGateForRepo");
+    expect(gateSource).toContain("export function resolveGateFromTrustedHumanForRepo");
+    expect(gateSource.split("\n").length).toBeLessThan(120);
   });
 
   it("does not expose trusted-human gate approval as a model tool", () => {
