+++ packages/pi-conductor/extensions/conductor.ts
@@ -1,11 +1,7 @@
 import { existsSync, rmSync } from "node:fs";
 import { resolve } from "node:path";
-import {
-  type ConductorBackendDispatcher,
-  type ConductorBackendsStatus,
-  getConductorBackendAdapter,
-  inspectConductorBackends,
-} from "./backends.js";
+import { type ConductorBackendDispatcher, getConductorBackendAdapter, inspectConductorBackends } from "./backends.js";
+import { createGateForRepo } from "./gate-service.js";
 import {
   commitAllChanges,
   createPullRequest,
@@ -13,8 +9,8 @@ import {
   validatePrPreconditions,
   validatePushPreconditions,
 } from "./git-pr.js";
-import { computeNextActions, incompleteDependenciesForTask } from "./next-actions.js";
-import { planObjectiveForRepo, refreshObjectiveStatusForRepo } from "./objective-service.js";
+import { computeNextActions } from "./next-actions.js";
+import { planObjectiveForRepo } from "./objective-service.js";
 import { getOrCreateRunForRepo, mutateRepoRunSync } from "./repo-run.js";
 import {
   createWorkerSessionRuntime,
@@ -25,28 +21,25 @@ import {
 import { type SchedulerFairness, type SchedulerPolicyName, selectSchedulerActions } from "./scheduler-selection.js";
 import {
   addConductorArtifact,
-  addTask,
   addWorker,
   appendConductorEvent,
-  assignTaskToWorker,
-  cancelTaskRun,
   completeTaskRun,
-  createConductorGate,
-  createTaskRecord,
   createWorkerRecord,
-  linkTaskToObjective,
   markConductorGateUsed,
   reconcileRunLeases,
-  recordTaskCompletion,
-  recordTaskProgress,
   removeWorker,
-  resolveConductorGate,
   setWorkerPrState,
-  startTaskRun,
-  updateTask,
 } from "./storage.js";
+import {
+  assignTaskForRepo,
+  createFollowUpTaskForRepo,
+  createTaskForRepo,
+  recordTaskCompletionForRepo,
+  recordTaskProgressForRepo,
+  retryTaskForRepo,
+  startTaskRunForRepo,
+} from "./task-service.js";
 import type {
-  ConductorActor,
   ConductorEventType,
   ConductorNextAction,
   ConductorNextActionsResponse,
@@ -54,13 +47,6 @@ import type {
   ConductorResourceRefs,
   ConductorResourceTimeline,
   ConductorTaskBrief,
-  EvidenceBundle,
-  EvidenceBundlePurpose,
-  GateRecord,
-  GateStatus,
-  ObjectiveRecord,
-  ReadinessCheck,
-  ReadinessPurpose,
   RunAttemptRecord,
   RunRecord,
   TaskContractInput,
@@ -69,6 +55,8 @@ import type {
   WorkerRunResult,
 } from "./types.js";
 
+export { buildEvidenceBundleForRepo, checkReadinessForRepo } from "./evidence-service.js";
+export { createGateForRepo, resolveGateForRepo, resolveGateFromTrustedHumanForRepo } from "./gate-service.js";
 export { computeNextActions } from "./next-actions.js";
 export {
   createObjectiveForRepo,
@@ -78,7 +66,19 @@ export {
   updateObjectiveForRepo,
 } from "./objective-service.js";
 export { getOrCreateRunForRepo } from "./repo-run.js";
+export { buildBlockingDiagnosisForRepo, prepareHumanReviewForRepo } from "./review-service.js";
 export type { SchedulerFairness, SchedulerPolicyName } from "./scheduler-selection.js";
+export {
+  assignTaskForRepo,
+  cancelTaskRunForRepo,
+  createFollowUpTaskForRepo,
+  createTaskForRepo,
+  recordTaskCompletionForRepo,
+  recordTaskProgressForRepo,
+  retryTaskForRepo,
+  startTaskRunForRepo,
+  updateTaskForRepo,
+} from "./task-service.js";
 
 function resolveSchedulerPolicy(input: { policy?: SchedulerPolicyName; executeRuns?: boolean }): {
   policy: SchedulerPolicyName;
@@ -438,109 +438,6 @@ export function buildProjectBriefForRepo(
   };
 }
 
-export function buildBlockingDiagnosisForRepo(
-  repoRoot: string,
-  input: { objectiveId?: string; taskId?: string } = {},
-): {
-  markdown: string;
-  blockers: Array<{
-    kind: string;
-    gateId?: string;
-    taskId?: string;
-    message: string;
-    nextToolCall: { name: string; params: Record<string, unknown> } | null;
-  }>;
-} {
-  const run = getOrCreateRunForRepo(repoRoot);
-  const taskIds = input.objectiveId
-    ? new Set(run.objectives.find((objective) => objective.objectiveId === input.objectiveId)?.taskIds ?? [])
-    : null;
-  const blockers = [] as Array<{
-    kind: string;
-    gateId?: string;
-    taskId?: string;
-    message: string;
-    nextToolCall: { name: string; params: Record<string, unknown> } | null;
-  }>;
-  for (const gate of run.gates.filter(
-    (entry) =>
-      entry.status === "open" &&
-      (!input.objectiveId ||
-        entry.resourceRefs.objectiveId === input.objectiveId ||
-        (entry.resourceRefs.taskId && taskIds?.has(entry.resourceRefs.taskId))) &&
-      (!input.taskId || entry.resourceRefs.taskId === input.taskId),
-  )) {
-    blockers.push({
-      kind: "gate",
-      gateId: gate.gateId,
-      message: gate.requestedDecision,
-      nextToolCall: ["needs_input", "needs_review"].includes(gate.type)
-        ? {
-            name: "conductor_resolve_gate",
-            params: { gateId: gate.gateId, status: "approved", resolutionReason: "<decision>", actorId: "parent" },
-          }
-        : null,
-    });
-  }
-  const tasks = input.taskId
-    ? run.tasks.filter((task) => task.taskId === input.taskId)
-    : input.objectiveId
-      ? run.tasks.filter((task) => taskIds?.has(task.taskId))
-      : run.tasks;
-  for (const task of tasks) {
-    for (const dependency of incompleteDependenciesForTask(run, task)) {
-      blockers.push({
-        kind: "dependency",
-        taskId: task.taskId,
-        message: `Task ${task.taskId} waits for ${dependency.taskId}`,
-        nextToolCall: { name: "conductor_task_brief", params: { taskId: dependency.taskId } },
-      });
-    }
-  }
-  const markdown = [
-    "# Conductor Blocking Diagnosis",
-    "",
-    blockers.length === 0
-      ? "- no blockers"
-      : blockers.map((blocker) => `- ${blocker.kind}: ${blocker.message}`).join("\n"),
-  ].join("\n");
-  return { markdown, blockers };
-}
-
-export function prepareHumanReviewForRepo(
-  repoRoot: string,
-  input: { objectiveId?: string; taskId?: string } = {},
-): {
-  markdown: string;
-  objective: ObjectiveRecord | null;
-  task: TaskRecord | null;
-  nextActions: ConductorNextAction[];
-  blockers: ReturnType<typeof buildBlockingDiagnosisForRepo>["blockers"];
-} {
-  const run = getOrCreateRunForRepo(repoRoot);
-  const objective = input.objectiveId
-    ? (run.objectives.find((entry) => entry.objectiveId === input.objectiveId) ?? null)
-    : null;
-  const task = input.taskId ? (run.tasks.find((entry) => entry.taskId === input.taskId) ?? null) : null;
-  const nextActions = computeNextActions(run, { objectiveId: input.objectiveId, maxActions: 5 }).actions;
-  const blockers = buildBlockingDiagnosisForRepo(repoRoot, input).blockers;
-  const markdown = [
-    "# Conductor Human Review Packet",
-    "",
-    objective
-      ? `Objective: ${objective.title} [${objective.objectiveId}] status=${objective.status}`
-      : "Objective: project",
-    task ? `Task: ${task.title} [${task.taskId}] state=${task.state}` : "Task: none selected",
-    "",
-    "## Blockers",
-    blockers.length === 0 ? "- none" : blockers.map((blocker) => `- ${blocker.kind}: ${blocker.message}`).join("\n"),
-    "",
-    "## Next Actions",
-    nextActions.length === 0 ? "- none" : nextActions.map((action) => `- ${action.kind}: ${action.title}`).join("\n"),
-  ].join("\n");
-  return { markdown, objective, task, nextActions, blockers };
-}
-
 async function executeConductorAction(
   repoRoot: string,
   action: ConductorNextAction,
@@ -816,162 +713,6 @@ export function getNextActionsForRepo(
   return computeNextActions(run, { ...input, reconciledPreview: input.reconcile ?? true });
 }
 
-function createTaskId(): string {
-  return `task-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-}
-
-function createRunId(): string {
-  return `run-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-}
-
-function leaseExpiryFromNow(leaseSeconds: number): string {
-  return new Date(Date.now() + leaseSeconds * 1000).toISOString();
-}
-
-export function createTaskForRepo(
-  repoRoot: string,
-  input: { title: string; prompt: string; objectiveId?: string; dependsOnTaskIds?: string[] },
-): TaskRecord {
-  let task!: TaskRecord;
-  mutateRepoRunSync(repoRoot, (run) => {
-    const missingDependency = input.dependsOnTaskIds?.find(
-      (taskId) => !run.tasks.some((entry) => entry.taskId === taskId),
-    );
-    if (missingDependency) {
-      throw new Error(`Task dependency ${missingDependency} not found`);
-    }
-    task = createTaskRecord({
-      taskId: createTaskId(),
-      title: input.title,
-      prompt: input.prompt,
-      objectiveId: input.objectiveId,
-      dependsOnTaskIds: input.dependsOnTaskIds,
-    });
-    return input.objectiveId
-      ? linkTaskToObjective(addTask(run, task), input.objectiveId, task.taskId)
-      : addTask(run, task);
-  });
-  return task;
-}
-
-export function updateTaskForRepo(
-  repoRoot: string,
-  input: { taskId: string; title?: string; prompt?: string },
-): TaskRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => updateTask(run, input));
-  const task = updatedRun.tasks.find((entry) => entry.taskId === input.taskId);
-  if (!task) {
-    throw new Error(`Task ${input.taskId} disappeared during update`);
-  }
-  return task;
-}
-
-export function assignTaskForRepo(repoRoot: string, taskId: string, workerId: string): TaskRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => assignTaskToWorker(run, taskId, workerId));
-  const task = updatedRun.tasks.find((entry) => entry.taskId === taskId);
-  if (!task) {
-    throw new Error(`Task ${taskId} disappeared during assignment`);
-  }
-  return task;
-}
-
-export function recordTaskProgressForRepo(
-  repoRoot: string,
-  input: {
-    runId: string;
-    taskId: string;
-    progress: string;
-    idempotencyKey?: string;
-    artifact?: {
-      type: "note" | "test_result" | "changed_files" | "log" | "completion_report" | "pr_evidence" | "other";
-      ref: string;
-      metadata?: Record<string, unknown>;
-    };
-  },
-): TaskRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => recordTaskProgress(run, input));
-  const task = updatedRun.tasks.find((entry) => entry.taskId === input.taskId);
-  if (!task) {
-    throw new Error(`Task ${input.taskId} disappeared during progress update`);
-  }
-  return task;
-}
-
-export function recordTaskCompletionForRepo(
-  repoRoot: string,
-  input: {
-    runId: string;
-    taskId: string;
-    status: "succeeded" | "partial" | "blocked" | "failed" | "aborted";
-    completionSummary: string;
-    idempotencyKey?: string;
-    artifact?: {
-      type: "note" | "test_result" | "changed_files" | "log" | "completion_report" | "pr_evidence" | "other";
-      ref: string;
-      metadata?: Record<string, unknown>;
-    };
-  },
-): TaskRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => recordTaskCompletion(run, input));
-  const task = updatedRun.tasks.find((entry) => entry.taskId === input.taskId);
-  if (!task) {
-    throw new Error(`Task ${input.taskId} disappeared during completion update`);
-  }
-  if (task.objectiveId) {
-    refreshObjectiveStatusForRepo(repoRoot, task.objectiveId);
-    return getOrCreateRunForRepo(repoRoot).tasks.find((entry) => entry.taskId === input.taskId) ?? task;
-  }
-  return task;
-}
-
-export function createGateForRepo(
-  repoRoot: string,
-  input: {
-    type: GateRecord["type"];
-    resourceRefs: ConductorResourceRefs;
-    requestedDecision: string;
-    gateId?: string;
-    operation?: GateRecord["operation"];
-    targetRevision?: number | null;
-    expiresAt?: string | null;
-  },
-): GateRecord {
-  const gateId = input.gateId ?? `gate-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => createConductorGate(run, { ...input, gateId }));
-  const gate = updatedRun.gates.find((entry) => entry.gateId === gateId);
-  if (!gate) {
-    throw new Error(`Gate ${gateId} disappeared during creation`);
-  }
-  return gate;
-}
-
-export function resolveGateForRepo(
-  repoRoot: string,
-  input: { gateId: string; status: Exclude<GateStatus, "open">; actor: ConductorActor; resolutionReason: string },
-): GateRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => resolveConductorGate(run, input));
-  const gate = updatedRun.gates.find((entry) => entry.gateId === input.gateId);
-  if (!gate) {
-    throw new Error(`Gate ${input.gateId} disappeared during resolution`);
-  }
-  return gate;
-}
-
-export function resolveGateFromTrustedHumanForRepo(
-  repoRoot: string,
-  input: { gateId: string; status: Exclude<GateStatus, "open">; humanId: string; resolutionReason: string },
-): GateRecord {
-  if (!input.humanId.trim()) {
-    throw new Error("Trusted human resolver requires a non-empty humanId");
-  }
-  return resolveGateForRepo(repoRoot, {
-    gateId: input.gateId,
-    status: input.status,
-    resolutionReason: input.resolutionReason,
-    actor: { type: "human", id: input.humanId },
-  });
-}
-
 export async function dispatchTaskRunForRepo(
   repoRoot: string,
   input: {
@@ -1064,135 +805,6 @@ export async function dispatchTaskRunForRepo(
   };
 }
 
-export function startTaskRunForRepo(
-  repoRoot: string,
-  input: {
-    taskId: string;
-    workerId?: string;
-    leaseSeconds?: number;
-    runId?: string;
-    backend?: RunAttemptRecord["backend"];
-    allowFollowUpTasks?: boolean;
-    inspectBackends?: () => ConductorBackendsStatus;
-  },
-): { run: RunAttemptRecord; taskContract: TaskContractInput } {
-  const run = getOrCreateRunForRepo(repoRoot);
-  const task = run.tasks.find((entry) => entry.taskId === input.taskId);
-  if (!task) {
-    throw new Error(`Task ${input.taskId} not found`);
-  }
-  const workerId = input.workerId ?? task.assignedWorkerId;
-  if (!workerId) {
-    throw new Error(`Task ${input.taskId} is not assigned to a worker`);
-  }
-  const backend = input.backend ?? "native";
-  if (backend === "pi-subagents") {
-    const status = (input.inspectBackends ?? inspectConductorBackends)().piSubagents;
-    mutateRepoRunSync(repoRoot, (latest) =>
-      appendConductorEvent(latest, {
-        actor: { type: "backend", id: "pi-subagents" },
-        type: "backend.unavailable",
-        resourceRefs: { projectKey: latest.projectKey, taskId: input.taskId, workerId },
-        payload: {
-          backend,
-          diagnostic: status.available
-            ? "pi-subagents dispatch adapter is not implemented yet"
-            : (status.diagnostic ?? "pi-subagents backend is unavailable"),
-        },
-      }),
-    );
-    throw new Error(
-      `pi-subagents backend unavailable: ${status.available ? "dispatch adapter is not implemented yet" : (status.diagnostic ?? "not available")}`,
-    );
-  }
-  const runId = input.runId ?? createRunId();
-  const updatedRun = mutateRepoRunSync(repoRoot, (latest) =>
-    startTaskRun(latest, {
-      runId,
-      taskId: input.taskId,
-      workerId,
-      backend,
-      leaseExpiresAt: leaseExpiryFromNow(input.leaseSeconds ?? 900),
-    }),
-  );
-  const started = updatedRun.runs.find((entry) => entry.runId === runId);
-  const updatedTask = updatedRun.tasks.find((entry) => entry.taskId === input.taskId);
-  if (!started || !updatedTask) {
-    throw new Error(`Run ${runId} disappeared during start`);
-  }
-  return {
-    run: started,
-    taskContract: {
-      taskId: updatedTask.taskId,
-      runId: started.runId,
-      taskRevision: started.taskRevision,
-      goal: updatedTask.prompt,
-      constraints: [],
-      explicitCompletionTools: true,
-      allowFollowUpTasks: input.allowFollowUpTasks ?? false,
-    },
-  };
-}
-
-export function cancelTaskRunForRepo(repoRoot: string, input: { runId: string; reason: string }): RunRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => cancelTaskRun(run, input));
-  const task = updatedRun.tasks.find((entry) => entry.runIds.includes(input.runId));
-  if (task?.objectiveId) {
-    refreshObjectiveStatusForRepo(repoRoot, task.objectiveId);
-    return getOrCreateRunForRepo(repoRoot);
-  }
-  return updatedRun;
-}
-
-export function retryTaskForRepo(
-  repoRoot: string,
-  input: { taskId: string; workerId?: string; leaseSeconds?: number },
-): { run: RunAttemptRecord; taskContract: TaskContractInput } {
-  const run = getOrCreateRunForRepo(repoRoot);
-  const task = run.tasks.find((entry) => entry.taskId === input.taskId);
-  if (!task) {
-    throw new Error(`Task ${input.taskId} not found`);
-  }
-  if (task.activeRunId) {
-    throw new Error(`Task ${input.taskId} already has an active run`);
-  }
-  if (!["blocked", "failed", "needs_review", "canceled"].includes(task.state)) {
-    throw new Error(`Task ${input.taskId} is ${task.state} and is not eligible for retry`);
-  }
-  const started = startTaskRunForRepo(repoRoot, input);
-  const refreshedTask = getOrCreateRunForRepo(repoRoot).tasks.find((entry) => entry.taskId === input.taskId);
-  if (refreshedTask?.objectiveId) {
-    refreshObjectiveStatusForRepo(repoRoot, refreshedTask.objectiveId);
-  }
-  return started;
-}
-
-export function createFollowUpTaskForRepo(
-  repoRoot: string,
-  input: { runId: string; taskId: string; title: string; prompt: string },
-): TaskRecord {
-  let task!: TaskRecord;
-  const withEvent = mutateRepoRunSync(repoRoot, (run) => {
-    const runAttempt = run.runs.find((entry) => entry.runId === input.runId);
-    if (!runAttempt || runAttempt.taskId !== input.taskId || runAttempt.finishedAt) {
-      throw new Error(`Run ${input.runId} is not an active run for task ${input.taskId}`);
-    }
-    task = createTaskRecord({ taskId: createTaskId(), title: input.title, prompt: input.prompt });
-    const withTask = addTask(run, task);
-    return appendConductorEvent(withTask, {
-      actor: { type: "child_run", id: input.runId },
-      type: "task.followup_created",
-      resourceRefs: { projectKey: run.projectKey, taskId: task.taskId, runId: input.runId },
-      payload: { parentTaskId: input.taskId, title: input.title },
-    });
-  });
-  const created = withEvent.tasks.find((entry) => entry.taskId === task.taskId);
-  if (!created) {
-    throw new Error(`Follow-up task ${task.taskId} disappeared during creation`);
-  }
-  return created;
-}
-
 export async function delegateTaskForRepo(
   repoRoot: string,
   input: { title: string; prompt: string; workerName: string; startRun?: boolean; leaseSeconds?: number },
@@ -1224,238 +836,6 @@ export async function delegateTaskForRepo(
     task: current.tasks.find((entry) => entry.taskId === assigned.taskId) ?? assigned,
     run: started.run,
     taskContract: started.taskContract,
-  };
-}
-
-function isTerminalStatus(status: string): boolean {
-  return ["succeeded", "partial", "blocked", "failed", "aborted", "stale", "interrupted", "unknown_dispatch"].includes(
-    status,
-  );
-}
-
-export function buildEvidenceBundleForRepo(
-  repoRoot: string,
-  input: {
-    workerId?: string;
-    workerName?: string;
-    objectiveId?: string;
-    taskId?: string;
-    runId?: string;
-    purpose?: EvidenceBundlePurpose;
-    includeEvents?: boolean;
-    persistArtifact?: boolean;
-  },
-): EvidenceBundle {
-  const run = getOrCreateRunForRepo(repoRoot);
-  const selectedRun = input.runId ? run.runs.find((entry) => entry.runId === input.runId) : null;
-  const objective = input.objectiveId
-    ? (run.objectives.find((entry) => entry.objectiveId === input.objectiveId) ?? null)
-    : null;
-  const task = input.taskId
-    ? run.tasks.find((entry) => entry.taskId === input.taskId)
-    : selectedRun
-      ? run.tasks.find((entry) => entry.taskId === selectedRun.taskId)
-      : null;
-  const worker = input.workerId
-    ? (run.workers.find((entry) => entry.workerId === input.workerId) ?? null)
-    : input.workerName
-      ? (run.workers.find((entry) => entry.name === input.workerName) ?? null)
-      : selectedRun
-        ? (run.workers.find((entry) => entry.workerId === selectedRun.workerId) ?? null)
-        : task?.assignedWorkerId
-          ? (run.workers.find((entry) => entry.workerId === task.assignedWorkerId) ?? null)
-          : null;
-  const taskIds = new Set<string>();
-  const runIds = new Set<string>();
-  if (objective) {
-    for (const taskId of objective.taskIds) taskIds.add(taskId);
-  }
-  if (task) taskIds.add(task.taskId);
-  if (selectedRun) runIds.add(selectedRun.runId);
-  if (worker) {
-    for (const entry of run.tasks.filter((candidate) => candidate.assignedWorkerId === worker.workerId)) {
-      taskIds.add(entry.taskId);
-    }
-  }
-  for (const entry of run.runs) {
-    if (taskIds.has(entry.taskId) || runIds.has(entry.runId)) {
-      taskIds.add(entry.taskId);
-      runIds.add(entry.runId);
-    }
-  }
-  const tasks = run.tasks.filter((entry) => taskIds.has(entry.taskId));
-  const runs = run.runs.filter((entry) => runIds.has(entry.runId) || taskIds.has(entry.taskId));
-  const gates = run.gates.filter(
-    (gate) =>
-      (worker && gate.resourceRefs.workerId === worker.workerId) ||
-      (objective && gate.resourceRefs.objectiveId === objective.objectiveId) ||
-      (gate.resourceRefs.taskId !== undefined && taskIds.has(gate.resourceRefs.taskId)) ||
-      (gate.resourceRefs.runId !== undefined && runIds.has(gate.resourceRefs.runId)),
-  );
-  const artifacts = run.artifacts.filter(
-    (artifact) =>
-      (worker && artifact.resourceRefs.workerId === worker.workerId) ||
-      (objective && artifact.resourceRefs.objectiveId === objective.objectiveId) ||
-      (artifact.resourceRefs.taskId !== undefined && taskIds.has(artifact.resourceRefs.taskId)) ||
-      (artifact.resourceRefs.runId !== undefined && runIds.has(artifact.resourceRefs.runId)),
-  );
-  const eventMatches = (refs: { workerId?: string; objectiveId?: string; taskId?: string; runId?: string }) =>
-    (worker && refs.workerId === worker.workerId) ||
-    (objective && refs.objectiveId === objective.objectiveId) ||
-    (refs.taskId !== undefined && taskIds.has(refs.taskId)) ||
-    (refs.runId !== undefined && runIds.has(refs.runId));
-  const bundle: EvidenceBundle = {
-    purpose: input.purpose ?? "task_review",
-    generatedAt: new Date().toISOString(),
-    resourceRefs: {
-      projectKey: run.projectKey,
-      objectiveId: objective?.objectiveId,
-      workerId: worker?.workerId,
-      taskId: task?.taskId,
-      runId: selectedRun?.runId,
-    },
-    objective,
-    worker,
-    tasks,
-    runs,
-    gates,
-    artifacts,
-    events: input.includeEvents ? run.events.filter((event) => eventMatches(event.resourceRefs)) : undefined,
-    pr: worker?.pr ?? null,
-    summary: {
-      taskCount: tasks.length,
-      runCount: runs.length,
-      openGateCount: gates.filter((gate) => gate.status === "open").length,
-      artifactCount: artifacts.length,
-      terminalRunCount: runs.filter((entry) => isTerminalStatus(entry.status)).length,
-      completedTaskCount: tasks.filter((entry) => entry.state === "completed").length,
-      needsReviewTaskCount: tasks.filter((entry) => entry.state === "needs_review").length,
-      blockedTaskCount: tasks.filter((entry) => entry.state === "blocked").length,
-      failedTaskCount: tasks.filter((entry) => entry.state === "failed").length,
-    },
-  };
-  if (input.persistArtifact) {
-    const withArtifact = mutateRepoRunSync(repoRoot, (latest) =>
-      addConductorArtifact(latest, {
-        type: "other",
-        ref: `evidence://${bundle.purpose}/${task?.taskId ?? worker?.workerId ?? objective?.objectiveId ?? selectedRun?.runId ?? latest.projectKey}/${Date.now().toString(36)}`,
-        resourceRefs: bundle.resourceRefs,
-        producer: { type: "parent_agent", id: "conductor" },
-        metadata: {
-          purpose: bundle.purpose,
-          taskIds: tasks.map((entry) => entry.taskId),
-          runIds: runs.map((entry) => entry.runId),
-          artifactIds: artifacts.map((entry) => entry.artifactId),
-          summary: bundle.summary,
-        },
-      }),
-    );
-    bundle.persistedArtifact = withArtifact.artifacts.at(-1);
-  }
-  return bundle;
-}
-
-export function checkReadinessForRepo(
-  repoRoot: string,
-  input: {
-    workerId?: string;
-    workerName?: string;
-    taskId?: string;
-    purpose: ReadinessPurpose;
-    requireCompletionReport?: boolean;
-    requireTestEvidence?: boolean;
-    requireNoOpenGates?: boolean;
-    requireCommit?: boolean;
-    requirePush?: boolean;
-    requireApprovedReadyForPrGate?: boolean;
-  },
-): ReadinessCheck {
-  const bundle = buildEvidenceBundleForRepo(repoRoot, { ...input, purpose: input.purpose });
-  const blockers: ReadinessCheck["blockers"] = [];
-  const warnings: ReadinessCheck["warnings"] = [];
-  if (input.purpose === "task_review") {
-    const task = bundle.tasks[0];
-    if (!task)
-      blockers.push({ code: "missing_task", message: "Task was not found", resourceRefs: bundle.resourceRefs });
-    if (task && !["completed", "needs_review"].includes(task.state)) {
-      blockers.push({
-        code: "task_not_terminal",
-        message: `Task is ${task.state}`,
-        resourceRefs: { taskId: task.taskId },
-      });
-    }
-    if (
-      (input.requireNoOpenGates ?? true) &&
-      bundle.gates.some((gate) => gate.status === "open" && gate.type !== "needs_review")
-    ) {
-      blockers.push({ code: "open_gate", message: "Open blocking gate exists", resourceRefs: bundle.resourceRefs });
-    }
-    if (
-      (input.requireCompletionReport ?? true) &&
-      !bundle.artifacts.some((artifact) => artifact.type === "completion_report")
-    ) {
-      blockers.push({
-        code: "missing_completion_report",
-        message: "Completion report artifact is required",
-        resourceRefs: bundle.resourceRefs,
-      });
-    }
-    if (input.requireTestEvidence && !bundle.artifacts.some((artifact) => artifact.type === "test_result")) {
-      blockers.push({
-        code: "missing_test_result",
-        message: "Test result artifact is required",
-        resourceRefs: bundle.resourceRefs,
-      });
-    }
-  } else {
-    if (!bundle.worker)
-      blockers.push({ code: "missing_worker", message: "Worker was not found", resourceRefs: bundle.resourceRefs });
-    if (bundle.tasks.filter((task) => ["completed", "needs_review"].includes(task.state)).length === 0) {
-      blockers.push({
-        code: "task_not_terminal",
-        message: "No completed or reviewable worker tasks found",
-        resourceRefs: bundle.resourceRefs,
-      });
-    }
-    if ((input.requireCommit ?? true) && !bundle.worker?.pr.commitSucceeded) {
-      blockers.push({
-        code: "missing_commit",
-        message: "Worker commit is required",
-        resourceRefs: bundle.resourceRefs,
-      });
-    }
-    if ((input.requirePush ?? true) && !bundle.worker?.pr.pushSucceeded) {
-      blockers.push({ code: "missing_push", message: "Worker push is required", resourceRefs: bundle.resourceRefs });
-    }
-    const readyGate = bundle.gates.find((gate) => gate.type === "ready_for_pr" && gate.status === "approved");
-    if ((input.requireApprovedReadyForPrGate ?? true) && !readyGate) {
-      blockers.push({
-        code: "missing_ready_for_pr_gate",
-        message: "Approved ready_for_pr gate is required",
-        resourceRefs: bundle.resourceRefs,
-      });
-    }
-    if (bundle.worker?.pr.url)
-      warnings.push({
-        code: "pr_already_created",
-        message: "Worker already has a PR",
-        resourceRefs: bundle.resourceRefs,
-      });
-  }
-  const status =
-    blockers.length === 0
-      ? "ready"
-      : blockers.some((blocker) => blocker.code === "open_gate")
-        ? "blocked"
-        : "not_ready";
-  return {
-    purpose: input.purpose,
-    status,
-    generatedAt: new Date().toISOString(),
-    resourceRefs: bundle.resourceRefs,
-    bundle,
-    blockers,
-    warnings,
   };
 }
 
@@ -1841,10 +1221,6 @@ function mapWorkerRunStatusToRunStatus(status: WorkerRunResult["status"]): "succ
   }
 }
 
-function createGateId(): string {
-  return `gate-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-}
-
 export async function runTaskForRepo(repoRoot: string, taskId: string, signal?: AbortSignal): Promise<WorkerRunResult> {
   const started = startTaskRunForRepo(repoRoot, { taskId });
   const currentRun = getOrCreateRunForRepo(repoRoot);
@@ -1904,7 +1280,6 @@ export async function runTaskForRepo(repoRoot: string, taskId: string, signal?: 
   });
   if (createdFallbackCompletion && runtimeResult.status === "success") {
     createGateForRepo(repoRoot, {
-      gateId: createGateId(),
       type: "needs_review",
       resourceRefs: { taskId, runId: started.run.runId, workerId: worker.workerId },
       requestedDecision: `Review task ${taskId}: native worker exited without explicit conductor_child_complete`,
