+++ packages/pi-conductor/extensions/storage.ts
@@ -17,6 +17,7 @@ import { deriveProjectKey } from "./project-key.js";
 import { defaultOperationForGate, normalizeProjectRecord } from "./storage-normalize.js";
 import { validateRunRecord } from "./storage-validation.js";
 
+export { queryConductorArtifacts, queryConductorEvents } from "./storage-query.js";
 export { validateRunRecord } from "./storage-validation.js";
 
 import type {
@@ -230,82 +231,6 @@ export function appendConductorEvent(
     revision: projectRevision,
     events: [...normalized.events, event],
     updatedAt: now,
-  };
-}
-
-export function queryConductorArtifacts(
-  run: RunRecord,
-  input: {
-    workerId?: string;
-    taskId?: string;
-    runId?: string;
-    gateId?: string;
-    artifactId?: string;
-    type?: ArtifactType;
-    afterIndex?: number;
-    limit?: number;
-  } = {},
-): { artifacts: ArtifactRecord[]; lastIndex: number | null; hasMore: boolean } {
-  const normalized = normalizeProjectRecord(run);
-  const limit = Math.max(1, Math.min(input.limit ?? 20, 100));
-  const filtered = normalized.artifacts
-    .map((artifact, index) => ({ artifact, index: index + 1 }))
-    .filter(({ artifact, index }) => {
-      if (input.afterIndex !== undefined && index <= input.afterIndex) {
-        return false;
-      }
-      if (input.type && artifact.type !== input.type) {
-        return false;
-      }
-      for (const key of ["workerId", "taskId", "runId", "gateId", "artifactId"] as const) {
-        if (input[key] && artifact.resourceRefs[key] !== input[key] && artifact.artifactId !== input[key]) {
-          return false;
-        }
-      }
-      return true;
-    });
-  const page = filtered.slice(0, limit);
-  return {
-    artifacts: page.map((entry) => entry.artifact),
-    lastIndex: page.at(-1)?.index ?? null,
-    hasMore: filtered.length > page.length,
-  };
-}
-
-export function queryConductorEvents(
-  run: RunRecord,
-  input: {
-    workerId?: string;
-    taskId?: string;
-    runId?: string;
-    gateId?: string;
-    artifactId?: string;
-    type?: string;
-    afterSequence?: number;
-    limit?: number;
-  } = {},
-): { events: ConductorEvent[]; lastSequence: number | null; hasMore: boolean } {
-  const normalized = normalizeProjectRecord(run);
-  const limit = Math.max(1, Math.min(input.limit ?? 20, 100));
-  const filtered = normalized.events.filter((event) => {
-    if (input.afterSequence !== undefined && event.sequence <= input.afterSequence) {
-      return false;
-    }
-    if (input.type && event.type !== input.type) {
-      return false;
-    }
-    for (const key of ["workerId", "taskId", "runId", "gateId", "artifactId"] as const) {
-      if (input[key] && event.resourceRefs[key] !== input[key]) {
-        return false;
-      }
-    }
-    return true;
-  });
-  const events = filtered.slice(0, limit);
-  return {
-    events,
-    lastSequence: events.at(-1)?.sequence ?? null,
-    hasMore: filtered.length > events.length,
   };
 }
